Add files via upload
This project analyzes customer demographics and purchasing behavior to understand factors influencing bike purchases. The dataset includes attributes such as age, income, education, occupation, and commute distance. Data is cleaned and categorized in the Working Sheet, summarized in the Pivot Table to explore income and purchasing trends, and visualized in the Dashboard for key insights.